# specs/001-mlops-platform/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuotaSet": {
        "properties": {
          "budget_gpu_seconds": {
            "minimum": 0.0,
            "title": "Budget Gpu Seconds",
            "type": "integer"
          },
          "window": {
            "description": "daily | weekly | monthly",
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "window",
          "budget_gpu_seconds"
        ],
        "title": "QuotaSet",
        "type": "object"
      },
...
      "Reorder": {
        "properties": {
          "position": {
            "description": "1-based target position in the queued lane",
            "minimum": 1.0,
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "position"
        ],
        "title": "Reorder",
        "type": "object"
      },
...
      "TenantCreate": {
        "properties": {
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TenantCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MLOps-Lite Gateway",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/jobs/{job_id}/{action}": {
      "post": {
        "description": "Pin/pause/resume/reorder a QUEUED job (T649).\n\nA **running** job is never touched by any of these \u2014 that is FR-010/FR-023a, and it is enforced\nat the agent rather than here so the rule holds for every caller, not just this route.",
        "operationId": "job_override_admin_jobs__job_id___action__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "title": "Action",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reorder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Override",
        "tags": [
          "broker-admin"
        ]
      }
    },
    "/admin/keys/{key_id}/revoke": {
      "post": {
        "operationId": "revoke_key_admin_keys__key_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Key",
        "tags": [
          "broker-admin"
        ]
      }
    },
    "/admin/queue": {
      "get": {
        "description": "Live queue, resident set, and both VRAM bounds' terms (T689, SC-006/SC-009).\n\nProxied from the host agent, which is the sole GPU-ordering authority \u2014 the gateway holds no\nadmission state of its own, and synthesizing any here would create a second, lagging answer to\na question with one authority.",
        "operationId": "queue_admin_queue_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue",
        "tags": [
          "broker-admin"
        ]
      }
    },
    "/admin/tenants": {
      "get": {
        "operationId": "list_tenants_admin_tenants_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tenants",
        "tags": [
          "broker-admin"
        ]
      },
      "post": {
        "description": "Create a tenant and its first key. The raw key is in this response and nowhere else.",
        "operationId": "create_tenant_admin_tenants_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tenant",
        "tags": [
          "broker-admin"
        ]
      }
    },
    "/admin/tenants/{tenant_id}/enable": {
      "post": {
        "operationId": "enable_tenant_admin_tenants__tenant_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Tenant",
        "tags": [
          "broker-admin"
        ]
      }
    },
    "/admin/tenants/{tenant_id}/keys": {
      "post": {
        "description": "Rotate (default) or add a key.\n\nRotation revokes every prior key; `?add=true` issues an additional one instead. Rotation is the\ndefault because it is what \"my key leaked\" needs, and an `add` that silently left the leaked key\nlive would be the wrong default for the dangerous case.",
        "operationId": "rotate_tenant_key_admin_tenants__tenant_id__keys_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "add",
            "required": false,
            "schema": {
              "default": false,
              "title": "Add",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Tenant Key",
        "tags": [
          "broker-admin"
        ]
      }
    },
    "/admin/tenants/{tenant_id}/quota": {
      "put": {
        "operationId": "set_quota_admin_tenants__tenant_id__quota_put",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotaSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Quota",
        "tags": [
          "broker-admin"
        ]
      }
    },
    "/admin/tenants/{tenant_id}/revoke": {
      "post": {
        "description": "Disable a tenant: every key is denied on the next request; in-flight work ends gracefully.\n\nThe keys themselves are untouched, so re-enabling restores the tenant's exact key set \u2014 a\ndisable that destroyed keys would make \"pause this tenant\" and \"expel this tenant\" the same\nirreversible action.",
        "operationId": "revoke_tenant_admin_tenants__tenant_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Tenant",
        "tags": [
          "broker-admin"
        ]
      }
    },
    "/admin/usage": {
      "get": {
        "description": "Per-tenant window consumption plus recent ledger rows (FR-017, SC-004).\n\n`consumed` counts settled ledger rows AND outstanding reservations, which is the number the\nquota is actually enforced against \u2014 reporting only the settled half would show an operator a\ntenant well inside its budget at the moment the broker refuses it.",
        "operationId": "usage_admin_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant",
            "required": false,
            "schema": {
              "title": "Tenant",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage",
        "tags": [
          "broker-admin"
        ]
      }
    },
...
    "/v1/usage": {
      "get": {
        "description": "A tenant's OWN usage. Scoped to the authenticated tenant by construction \u2014 there is no path\nparameter to confuse, which is the simplest possible answer to the cross-tenant question (T633).",
        "operationId": "my_usage_v1_usage_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Usage",
        "tags": [
          "broker-inference"
        ]
      }
    },
...
  }
}